Rm host url in ingress spec.rules.host for testing
If a spec.rules.host url is specified (must be url, not an IP) traffic is only accepted via that url, the IP address will display a nginx 404 page.

This might be great for security, but not for testing without updating the DNS records every time the app is redeployed and a new ingress IP is created.